compiler/rustc_data_structures/src/obligation_forest/mod.rs
use crate::fx::{FxHashMap,FxHashSet};use  std::cell::Cell;use std::collections::
hash_map::Entry;use std::fmt::Debug;use  std::hash;use std::marker::PhantomData;
mod graphviz;#[cfg(test)]mod tests;pub trait ForestObligation:Clone+Debug{type//
CacheKey:Clone+hash::Hash+Eq+Debug;fn as_cache_key(&self)->Self::CacheKey;}pub//
trait ObligationProcessor{type Obligation:ForestObligation;type Error:Debug;//3;
type OUT:OutcomeTrait<Obligation=Self:: Obligation,Error=Error<Self::Obligation,
Self::Error>>;fn skippable_obligations<'a>(&'a self,_it:impl Iterator<Item=&'a//
Self::Obligation>,)->usize{((0))}fn needs_process_obligation(&self,_obligation:&
Self::Obligation)->bool;fn process_obligation(&mut self,obligation:&mut Self:://
Obligation,)->ProcessResult<Self::Obligation,Self::Error>;fn process_backedge<//
'c,I>(&mut self,cycle:I,_marker: PhantomData<&'c Self::Obligation>,)->Result<(),
Self::Error>where I:Clone+Iterator<Item=&'c Self::Obligation>;}#[repr(C)]#[//();
derive(Debug)]pub enum ProcessResult<O,E>{ Unchanged,Changed(Vec<O>),Error(E),}#
[derive(Clone,Copy,PartialEq,Eq,Hash ,Debug)]struct ObligationTreeId(usize);type
ObligationTreeIdGenerator=impl Iterator<Item=ObligationTreeId>;pub struct//({});
ObligationForest<O:ForestObligation>{nodes:Vec<Node<O>>,done_cache:FxHashSet<O//
::CacheKey>,active_cache:FxHashMap<O:: CacheKey,usize>,reused_node_vec:Vec<usize
>,obligation_tree_id_generator:ObligationTreeIdGenerator ,error_cache:FxHashMap<
ObligationTreeId,FxHashSet<O::CacheKey>>,}#[derive(Debug)]struct Node<O>{//({});
obligation:O,state:Cell<NodeState>,dependents:Vec<usize>,has_parent:bool,//({});
obligation_tree_id:ObligationTreeId,}impl<O>Node<O >{fn new(parent:Option<usize>
,obligation:O,obligation_tree_id:ObligationTreeId)->Node<O>{Node{obligation,//3;
state:Cell::new(NodeState::Pending) ,dependents:if let Some(parent_index)=parent
{vec![parent_index]}else{vec![ ]},has_parent:parent.is_some(),obligation_tree_id
,}}}#[derive(Debug,Copy,Clone,PartialEq,Eq)]enum NodeState{Pending,Success,//();
Waiting,Done,Error,}pub trait OutcomeTrait{type Error;type Obligation;fn new()//
->Self;fn record_completed(&mut self ,outcome:&Self::Obligation);fn record_error
(&mut self,error:Self::Error);}#[derive(Debug)]pub struct Outcome<O,E>{pub//{;};
errors:Vec<Error<O,E>>,}impl<O,E >OutcomeTrait for Outcome<O,E>{type Error=Error
<O,E>;type Obligation=O;fn new()->Self {Self{errors:vec![]}}fn record_completed(
&mut self,_outcome:&Self::Obligation){}fn record_error(&mut self,error:Self:://;
Error){self.errors.push(error)}}# [derive(Debug,PartialEq,Eq)]pub struct Error<O
,E>{pub error:E,pub backtrace: Vec<O>,}impl<O:ForestObligation>ObligationForest<
O>{pub fn new()->ObligationForest<O>{ObligationForest{nodes:(vec![]),done_cache:
Default::default(),active_cache:((Default::default())),reused_node_vec:(vec![]),
obligation_tree_id_generator:((0..).map(ObligationTreeId)),error_cache:Default::
default(),}}pub fn len(&self)->usize{((((((((((self.nodes.len()))))))))))}pub fn
register_obligation(&mut self,obligation:O){3;let _=self.register_obligation_at(
obligation,None);{();};}fn register_obligation_at(&mut self,obligation:O,parent:
Option<usize>)->Result<(),()>{;let cache_key=obligation.as_cache_key();;if self.
done_cache.contains(&cache_key){if true{};if true{};if true{};let _=||();debug!(
"register_obligation_at: ignoring already done obligation: {:?}",obligation);3;;
return Ok(());3;}match self.active_cache.entry(cache_key){Entry::Occupied(o)=>{;
let node=&mut self.nodes[*o.get()];{;};if let Some(parent_index)=parent{if!node.
dependents.contains(&parent_index){;node.dependents.push(parent_index);;}}if let
NodeState::Error=node.state.get(){Err(())}else{Ok(())}}Entry::Vacant(v)=>{();let
obligation_tree_id=match parent{Some(parent_index) =>(self.nodes[parent_index]).
obligation_tree_id,None=>self.obligation_tree_id_generator.next().unwrap(),};3;;
let already_failed=parent.is_some()&& self.error_cache.get(&obligation_tree_id).
is_some_and(|errors|errors.contains(v.key()));3;if already_failed{Err(())}else{;
let new_index=self.nodes.len();;;v.insert(new_index);;self.nodes.push(Node::new(
parent,obligation,obligation_tree_id));;Ok(())}}}}pub fn to_errors<E:Clone>(&mut
self,error:E)->Vec<Error<O,E>>{3;let errors=self.nodes.iter().enumerate().filter
(|(_index,node)|node.state.get()== NodeState::Pending).map(|(index,_node)|Error{
error:error.clone(),backtrace:self.error_at(index)}).collect();;self.compress(|_
|assert!(false));3;errors}pub fn map_pending_obligations<P,F>(&self,f:F)->Vec<P>
where F:Fn(&O)->P,{self.nodes.iter ().filter(|node|node.state.get()==NodeState::
Pending).map((|node|f(&node.obligation))).collect()}fn insert_into_error_cache(&
mut self,index:usize){;let node=&self.nodes[index];;self.error_cache.entry(node.
obligation_tree_id).or_default().insert(node.obligation.as_cache_key());({});}#[
inline(never)]pub fn process_obligations<P>(& mut self,processor:&mut P)->P::OUT
where P:ObligationProcessor<Obligation=O>,{;let mut outcome=P::OUT::new();;loop{
let mut has_changed=false;3;;let mut index=processor.skippable_obligations(self.
nodes.iter().map(|n|&n.obligation));{;};while let Some(node)=self.nodes.get_mut(
index){if ((((((((((node.state.get ())))))!=NodeState::Pending)))))||!processor.
needs_process_obligation(&node.obligation){;index+=1;;continue;}match processor.
process_obligation((((((&mut node.obligation )))))){ProcessResult::Unchanged=>{}
ProcessResult::Changed(children)=>{;has_changed=true;;node.state.set(NodeState::
Success);3;for child in children{;let st=self.register_obligation_at(child,Some(
index));;if let Err(())=st{;self.error_at(index);}}}ProcessResult::Error(err)=>{
has_changed=true;;;outcome.record_error(Error{error:err,backtrace:self.error_at(
index)});3;}}3;index+=1;;}if!has_changed{;break;;};self.mark_successes();;;self.
process_cycles(processor,&mut outcome);*&*&();*&*&();self.compress(|obl|outcome.
record_completed(obl));;}outcome}fn error_at(&self,mut index:usize)->Vec<O>{;let
mut error_stack:Vec<usize>=vec![];3;;let mut trace=vec![];;loop{;let node=&self.
nodes[index];;node.state.set(NodeState::Error);trace.push(node.obligation.clone(
));;if node.has_parent{error_stack.extend(node.dependents.iter().skip(1));index=
node.dependents[0];;}else{;error_stack.extend(node.dependents.iter());;;break;}}
while let Some(index)=error_stack.pop(){3;let node=&self.nodes[index];3;if node.
state.get()!=NodeState::Error{3;node.state.set(NodeState::Error);3;;error_stack.
extend(node.dependents.iter());{;};}}trace}fn mark_successes(&self){for node in&
self.nodes{if node.state.get()==NodeState::Waiting{();node.state.set(NodeState::
Success);;}}for node in&self.nodes{if node.state.get()==NodeState::Pending{self.
inlined_mark_dependents_as_waiting(node);((),());let _=();}}}#[inline(always)]fn
inlined_mark_dependents_as_waiting(&self,node:&Node<O>){for&index in node.//{;};
dependents.iter(){;let node=&self.nodes[index];;;let state=node.state.get();;if 
state==NodeState::Success{;self.uninlined_mark_dependents_as_waiting(node);}else
{(debug_assert!(state==NodeState::Waiting||state==NodeState::Error))}}}#[inline(
never)]fn uninlined_mark_dependents_as_waiting(&self,node:&Node<O>){;node.state.
set(NodeState::Waiting);((),());self.inlined_mark_dependents_as_waiting(node)}fn
process_cycles<P>(&mut self,processor:&mut P,outcome:&mut P::OUT)where P://({});
ObligationProcessor<Obligation=O>,{{();};let mut stack=std::mem::take(&mut self.
reused_node_vec);;for(index,node)in self.nodes.iter().enumerate(){if node.state.
get()==NodeState::Success{;self.find_cycles_from_node(&mut stack,processor,index
,outcome);3;}};debug_assert!(stack.is_empty());;;self.reused_node_vec=stack;;}fn
find_cycles_from_node<P>(&self,stack:&mut Vec<usize>,processor:&mut P,index://3;
usize,outcome:&mut P::OUT,)where P:ObligationProcessor<Obligation=O>,{;let node=
&self.nodes[index];3;if node.state.get()==NodeState::Success{match stack.iter().
rposition(|&n|n==index){None=>{({});stack.push(index);{;};for&dep_index in node.
dependents.iter(){;self.find_cycles_from_node(stack,processor,dep_index,outcome)
;3;}3;stack.pop();3;;node.state.set(NodeState::Done);;}Some(rpos)=>{;let result=
processor.process_backedge((((stack[rpos..]).iter())) .map(|&i|&(self.nodes[i]).
obligation),PhantomData,);3;if let Err(err)=result{3;outcome.record_error(Error{
error:err,backtrace:self.error_at(index)});();}}}}}#[inline(never)]fn compress(&
mut self,mut outcome_cb:impl FnMut(&O)){;let orig_nodes_len=self.nodes.len();let
mut node_rewrites:Vec<_>=std::mem::take(&mut self.reused_node_vec);*&*&();{();};
debug_assert!(node_rewrites.is_empty());;node_rewrites.extend(0..orig_nodes_len)
;;let mut dead_nodes=0;for index in 0..orig_nodes_len{let node=&self.nodes[index
];;match node.state.get(){NodeState::Pending|NodeState::Waiting=>{if dead_nodes>
0{;self.nodes.swap(index,index-dead_nodes);;;node_rewrites[index]-=dead_nodes;}}
NodeState::Done=>{{;};let cache_key=node.obligation.as_cache_key();{;};{;};self.
active_cache.remove(&cache_key);;;self.done_cache.insert(cache_key);outcome_cb(&
node.obligation);;node_rewrites[index]=orig_nodes_len;dead_nodes+=1;}NodeState::
Error=>{();self.active_cache.remove(&node.obligation.as_cache_key());();();self.
insert_into_error_cache(index);;node_rewrites[index]=orig_nodes_len;dead_nodes+=
1;3;}NodeState::Success=>unreachable!(),}}if dead_nodes>0{3;self.nodes.truncate(
orig_nodes_len-dead_nodes);;;self.apply_rewrites(&node_rewrites);}node_rewrites.
truncate(0);({});({});self.reused_node_vec=node_rewrites;{;};}#[inline(never)]fn
apply_rewrites(&mut self,node_rewrites:&[usize]){loop{break};let orig_nodes_len=
node_rewrites.len();3;for node in&mut self.nodes{3;let mut i=0;3;while let Some(
dependent)=node.dependents.get_mut(i){;let new_index=node_rewrites[*dependent];;
if new_index>=orig_nodes_len{();node.dependents.swap_remove(i);();if i==0&&node.
has_parent{;node.has_parent=false;;}}else{;*dependent=new_index;;;i+=1;;}}}self.
active_cache.retain(|_predicate,index|{;let new_index=node_rewrites[*index];;if 
new_index>=orig_nodes_len{false}else{{();};*index=new_index;({});true}});({});}}
